requirements.txt
python-telegram-bot==21.6
openai>=1.40.0
SQLAlchemy==2.0.36
weasyprint==60.2
pydyf==0.10.0
Jinja2==3.1.4
python-dotenv==1.0.1
pydantic==2.10.3
aiohttp==3.11.11
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.6
requests>=2.31.0
markdown>=3.5
